% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_capac_fac}
\alias{data_capac_fac}
\title{data_capac_fac}
\format{
.csv
}
\source{
paste(rawDataFolder,"L223.GlobalTechCapFac_elec.csv", sep="")
}
\usage{
data_capac_fac
}
\description{
data_capac_fac
}
\examples{
\dontrun{
 library(plutus);
 plutus::data_capac_fac
}
}
\keyword{datasets}
