% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamInvest.R
\name{gcamInvest}
\alias{gcamInvest}
\title{gcamInvest}
\usage{
gcamInvest(
  gcamdatabase = NULL,
  queryFile = NULL,
  reReadData = T,
  dataProjFile = paste(getwd(), "/outputs/dataProj.proj", sep = ""),
  gcamdataFile = NULL,
  scenOrigNames = "All",
  scenNewNames = NULL,
  regionsSelect = NULL,
  dirOutputs = paste(getwd(), "/outputs", sep = ""),
  folderName = NULL,
  nameAppend = "",
  saveData = T
)
}
\arguments{
\item{gcamdatabase}{Default = NULL. Full path to GCAM database folder.}

\item{queryFile}{Defualt = NULL. When NULL plutus loads pre-saved xml file plutus::xmlElecQueries}

\item{reReadData}{If TRUE will read the GCAM data base and create a queryData.proj file
in the same folder as the GCAM database. If FALSE will load a '.proj' file if a file
with full path is provided otherwise it will search for a dataProj.proj file in the existing
folder which may have been created from an old run.}

\item{dataProjFile}{Default = NULL. Optional. A default 'dataProj.proj' is produced if no .Proj file is specified.}

\item{gcamdataFile}{Default = NULL. Optional. For example, gcamdataFile = "~/gcam-core-gcam-v5.3/input/gcamdata".
Use full path to GCAM 'gcamdata' folder that contains costs and capacity data. Data files including:

A23.globaltech_retirement.csv
L223.StubTechCapFactor_elec.csv
L223.GlobalTechCapFac_elec.csv
L2233.GlobalIntTechCapital_elec.csv
L2233.GlobalIntTechCapital_elec_cool.csv
L2233.GlobalTechCapital_elec_cool.csv
L2233.GlobalTechCapital_elecPassthru.csv}

\item{scenOrigNames}{Default = "All". Original Scenarios names in GCAM database in a string vector.
For example c('scenario1','scenario2).}

\item{scenNewNames}{New Names which may be shorter and more useful for figures etc.
Default will use Original Names. For example c('scenario1','scenario2)}

\item{regionsSelect}{The regions to analyze in a vector. Example c('Colombia','Argentina'). Full list:

USA, Africa_Eastern, Africa_Northern, Africa_Southern, Africa_Western, Australia_NZ, Brazil, Canada
Central America and Caribbean, Central Asia, China, EU-12, EU-15, Europe_Eastern, Europe_Non_EU,
European Free Trade Association, India, Indonesia, Japan, Mexico, Middle East, Pakistan, Russia,
South Africa, South America_Northern, South America_Southern, South Asia, South Korea, Southeast Asia,
Taiwan, Argentina, Colombia, Uruguay}

\item{dirOutputs}{Full path to directory for outputs}

\item{folderName}{Default = NULL}

\item{nameAppend}{Default = "". Name to append to saved files.}

\item{saveData}{Default = "T". Set to F if do not want to save any data to file.}
}
\value{
Returns (1) annual and cumulative costs and power generation of
premature retired electricity infrustructure (including hydropower);
(2) annual and cumulative costs and power generation of new capacity from new electricity infrustructures (including hydropower).
}
\description{
This function connects and reads a gcamdatabase or a query data proj file and calculates
electricity and hydropower investments.
}
\keyword{database,}
\keyword{gcam}
\keyword{gcam,}
\keyword{query}
