% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroInvest.R
\name{hydroInvest}
\alias{hydroInvest}
\title{hydroInvest}
\usage{
hydroInvest(addition_costs, start_year = 2010)
}
\arguments{
\item{addition_costs}{list formatted as produced by elecInvest.R function}

\item{start_year}{start year for analysis}
}
\value{
Returns data in a form required by gcamInvest.R
}
\description{
Function that calculates electricity subsector investment requirements from a given GCAM run.
}
\keyword{infrastructure}
\keyword{investments,}
